Implement TCO to avoid stack ovverflow
There is a recursion loop between stateSelfClosingStartTag and
stateSelfClosingStartTag which in the worst case consumes 2 stack
frames per input character.

We avoid this recursion by manual implementing tail call optimization
(TCO) which a go compiler does not do automatically.

Also add a unit test to reproduce the problem in the first place.

This fixes #16